+++ 002/utility_functions/UtilityFunctions.cpp
@@ -1,3 +1,16 @@
+#include <iostream>
+#include <vector>
 #include "UtilityFunctions.h"
 
-// Implement your functions here. Don't forget function comments!
+using namespace std;
+
+string Join(vector<string> pieces, string glue) {
+  string s;
+  for (int i = 0; i < pieces.size(); i++){
+    if(i < pieces.size() - 1)
+      s = s + pieces[i] + glue;
+    else
+      s = s + pieces[i];
+  };
+  return s;
+}
