+++ 002/utility_functions/main.cpp
@@ -1,10 +1,12 @@
-#include "UtilityFunctions.h"
+#include "UtilityFunctions.cpp"
+#include <iostream>
+#include <vector>
 
+using namespace std;
 
 int main() {
-    // the main function should have demonstrations
-    // of running your functions.
-    // make sure that the output is easily interpretable.
-    // If you'd like to define helper functions in this file,
-    // you are welcome to do so!
+    std::vector<string> my_fav = {"I", "Like", "PB&J"};
+    string glue = " ";
+    cout << Join(my_fav, glue) << '\n';
+    //Prints string "I Like PB&J"
 }
