+++ 002/utility_functions/test.cpp
@@ -7,17 +7,6 @@
 
 using namespace std;
 
-string Join(vector<string> pieces, string glue) {
-  string s;
-  for (int i = 0; i < pieces.size(); i++){
-    if(i < pieces.size() - 1)
-      s = s + pieces[i] + glue;
-    else
-      s = s + pieces[i];
-  };
-  return s;
-}
-
 TEST_CASE( "combines vector to string with glue in between", "[vector]" ) {
   std::vector<string> a{"wow"};
   string glue = "-";
